+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+# [3.0.0](https://github.com/AlexKaikin/evo-place-frontend/compare/v2.5.0...v3.0.0) (2024-02-22)
+
+
+* Merge pull request #69 from AlexKaikin/feature/220224-improvement-profile-to-club ([0717ee1](https://github.com/AlexKaikin/evo-place-frontend/commit/0717ee1dc59f7700ada7235078b5ed40c3d325d2)), closes [#69](https://github.com/AlexKaikin/evo-place-frontend/issues/69)
+
+
+### BREAKING CHANGES
+
+* Improvement user profile in club
+
+# [2.5.0](https://github.com/AlexKaikin/evo-place-frontend/compare/v2.4.1...v2.5.0) (2024-02-20)
+
+
+### Features
+
+* shop filter improvement ([8ae4345](https://github.com/AlexKaikin/evo-place-frontend/commit/8ae43453f9e12f5e45edcfbee18e5d8d6de1859a))
+
+## [2.4.1](https://github.com/AlexKaikin/evo-place-frontend/compare/v2.4.0...v2.4.1) (2024-02-18)
+
+
+### Performance Improvements
+
+* added profile settings to club ([3f0790d](https://github.com/AlexKaikin/evo-place-frontend/commit/3f0790d102a3ffb8e31f3605993669f5c4508d60))
+
+# [2.4.0](https://github.com/AlexKaikin/evo-place-frontend/compare/v2.3.1...v2.4.0) (2024-02-17)
+
+
+### Features
+
+* added subscriptions to club ([dade5d9](https://github.com/AlexKaikin/evo-place-frontend/commit/dade5d9e0957750c32bc63a118e7c3ca1020fb09))
+
 ## [2.3.1](https://github.com/AlexKaikin/evo-place-frontend/compare/v2.3.0...v2.3.1) (2024-02-11)